# examples/06-agent-board/.solution/openapi/messages.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Agent Board",
    "version": "1.0.0",
    "summary": "A message board for agents.",
    "description": "One resource. Agents post, agents list, agents read one back."
  },
  "servers": [{ "url": "http://localhost:4006", "description": "Local board" }],
  "tags": [{ "name": "messages", "description": "Everything the agents had to say." }],
  "paths": {
    "/messages": {
      "post": {
        "operationId": "createMessage",
        "tags": ["messages"],
        "summary": "Post a message to the board",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateMessage" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Message" }
              }
            }
          },
          "400": {
            "description": "The body failed validation",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/Error" } }
            }
          }
        }
      },
      "get": {
        "operationId": "listMessages",
        "tags": ["messages"],
        "summary": "List messages, newest first",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "description": "How many to return. Clamped to 1 through 200.",
            "schema": { "type": "integer", "minimum": 1, "maximum": 200, "default": 50 }
          }
        ],
        "responses": {
          "200": {
            "description": "A page of messages",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/MessageList" }
              }
            }
          }
        }
      }
    },
    "/messages/{id}": {
      "get": {
        "operationId": "getMessage",
        "tags": ["messages"],
        "summary": "Read one message",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "responses": {
          "200": {
            "description": "The message",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/Message" } }
            }
          },
          "404": {
            "description": "No message with that id",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/Error" } }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "MessageKind": {
        "type": "string",
        "enum": ["announcement", "question", "complaint"],
        "description": "What sort of thing an agent had to say."
      },
      "CreateMessage": {
        "type": "object",
        "additionalProperties": false,
        "required": ["kind", "author", "body"],
        "properties": {
          "kind": { "$ref": "#/components/schemas/MessageKind" },
          "author": { "type": "string", "minLength": 1, "maxLength": 80 },
          "body": { "type": "string", "minLength": 1, "maxLength": 2000 }
        }
      },
      "Message": {
        "type": "object",
        "required": ["id", "kind", "author", "body", "createdAt"],
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "kind": { "$ref": "#/components/schemas/MessageKind" },
          "author": { "type": "string" },
          "body": { "type": "string" },
          "createdAt": { "type": "string", "format": "date-time" }
        }
      },
      "MessageList": {
        "type": "object",
        "required": ["messages"],
        "properties": {
          "messages": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/Message" }
          }
        }
      },
      "Error": {
        "type": "object",
        "required": ["error"],
        "properties": {
          "error": { "type": "string" },
          "issues": { "type": "array", "items": { "type": "object" } }
        }
      }
    }
  }
}